+++ src/json/index.ts
@@ -1,5 +1,5 @@
 import { array } from "./array.js";
-import { json, stringifiedJson } from "./json.js";
+import { json } from "./json.js";
 import { object } from "./object.js";
 import { primitive } from "./primitive.js";
 
@@ -7,7 +7,6 @@ const module = Object.assign(json, {
   array,
   object,
   primitive,
-  stringified: stringifiedJson,
 });
 
 export { module as json };