+++ src/string/index.ts
@@ -1,8 +1,10 @@
 import { integer } from "./integer.js";
+import { json } from "./json.js";
 import { url } from "./url.js";
 
 const string = {
   integer,
+  json,
   url,
 };