+++ test/unit/string/index.ts
@@ -1,9 +1,11 @@
 import { injectIntegerTests } from "./integer.js";
+import { injectJsonTests } from "./json.js";
 import { injectUrlTests } from "./url.js";
 
 export function injectStringTests() {
   describe("string", () => {
     injectIntegerTests();
+    injectJsonTests();
     injectUrlTests();
   });
 }